fitness-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
email-validator==2.1.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
anthropic>=0.40.0
rapidfuzz>=3.0.0
